+++ Interface/AddOns/NDui/Core/Database.lua
@@ -45,6 +45,24 @@ if DB.isWW then
 	GetSpellCount = C_Spell.GetSpellCastCount or GetSpellCount
 	IsUsableSpell = C_Spell.IsSpellUsable or IsUsableSpell
 	IsCurrentSpell = C_Spell.IsCurrentSpell or IsCurrentSpell
+
+	local function EasyMenu_Initialize( frame, level, menuList )
+		for index = 1, #menuList do
+			local value = menuList[index]
+			if (value.text) then
+				value.index = index;
+				UIDropDownMenu_AddButton( value, level );
+			end
+		end
+	end
+	
+	function EasyMenu(menuList, menuFrame, anchor, x, y, displayMode, autoHideDelay )
+		if ( displayMode == "MENU" ) then
+			menuFrame.displayMode = displayMode;
+		end
+		UIDropDownMenu_Initialize(menuFrame, EasyMenu_Initialize, displayMode, nil, menuList);
+		ToggleDropDownMenu(1, nil, menuFrame, anchor, x, y, menuList, nil, autoHideDelay);
+	end
 end
 
 -- Colors
@@ -108,6 +126,8 @@ DB.ArrowUp = Media.."Hutu\\arrow"
 DB.afdianTex = Media.."Hutu\\Afdian"
 DB.patreonTex = Media.."Hutu\\Patreon"
 DB.sponsorTex = Media.."Hutu\\Sponsor"
+DB.curseforgeTex = Media.."Hutu\\CURSEFORGE.PNG"
+DB.boxTex = Media.."Hutu\\Box.PNG"
 DB.mailTex = "Interface\\Minimap\\Tracking\\Mailbox"
 DB.gearTex = "Interface\\WorldMap\\Gear_64"
 DB.eyeTex = "Interface\\Minimap\\Raid_Icon"		-- blue: \\Dungeon_Icon