+++ Interface/AddOns/NDui/Modules/Skins/Blizzard/FrameXML/QuestMapFrame.lua
@@ -140,7 +140,7 @@ tinsert(C.defaultThemes, function()
 		for button in QuestScrollFrame.titleFramePool:EnumerateActive() do
 			if not button.styled then
 				if button.Checkbox then
-					button.Checkbox:DisableDrawLayer("BACKGROUND")
+					B.StripTextures(button.Checkbox, 2)
 					B.CreateBDFrame(button.Checkbox, 0, true)
 				end
 				if button.Check then -- isWW removed?
@@ -155,7 +155,16 @@ tinsert(C.defaultThemes, function()
 		end
 
 		for header in QuestScrollFrame.campaignHeaderMinimalFramePool:EnumerateActive() do
-			ReskinQuestHeader(header)
+			if DB.isWW then
+				if header.CollapseButton and not header.styled then
+					B.StripTextures(header)
+					B.CreateBDFrame(header.Background, .25)
+					header.Highlight:SetColorTexture(1, 1, 1, .25)
+					header.styled = true
+				end
+			else
+				ReskinQuestHeader(header)
+			end
 		end
 
 		for header in QuestScrollFrame.covenantCallingsHeaderFramePool:EnumerateActive() do
@@ -170,6 +179,16 @@ tinsert(C.defaultThemes, function()
 		B.StripTextures(CompleteQuestFrame)
 		B.StripTextures(CompleteQuestFrame.CompleteButton)
 		B.Reskin(CompleteQuestFrame.CompleteButton)
+	end
+
+	-- Map legend
+	local mapLegend = QuestMapFrame.MapLegend
+	if mapLegend then
+		B.StripTextures(mapLegend.BorderFrame)
+		B.Reskin(mapLegend.BackButton)
+		B.ReskinTrimScroll(mapLegend.ScrollFrame.ScrollBar)
+		B.StripTextures(mapLegend.ScrollFrame)
+		B.CreateBDFrame(mapLegend.ScrollFrame, .25)
 	end
 
 	-- [[ Quest log popup detail frame ]]